set up bar graph to display merchants or categories based off filter and moved data processing functions for line and bar charts to utils helper file.
