+++ public/js/components/BarGraph.react.js
@@ -58,7 +58,7 @@ var BarGraph = React.createClass({
     return (
       <div id="bar-graph">
         <div className="graph-header">
-          <h2>Top Merchants</h2>
+          <h2>Top {GraphDataStore.getFilterValue().secondary}</h2>
         </div>
         <div id="chart_3"></div>
       </div>
